authentication/requirements.txt
argon2-cffi
fastapi
uvicorn
pydantic
authlib
pydantic-settings
pydantic[email]
passlib
python-dotenv
boto3
botocore
pyjwt
python-multipart
python-jose
oauthlib
httpx
itsdangerous
structlog
pytest
